# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "ASE Project API"
  version: "1.0.0"
  description: "API documentation for ASE Project"
servers:
  - url: "http://localhost:5050"
paths:
...
  /gacha:
    get:
      summary: "Get all gachas"
      tags: ["Gacha"]
      responses:
        '200':
          description: "List of all gachas"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gacha'
        '404':
          description: "Gachas not found"
    post:
      summary: "Create a new gacha"
      tags: ["Gacha"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGachaRequest'
      responses:
        '200':
          description: "Gacha created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  gachaId:
                    type: integer
                    description: "ID of the newly created gacha"
        '400':
          description: "Invalid gacha data"
  /gacha/{gachaId}:
    get:
      summary: "Get a single gacha by ID"
      tags: ["Gacha"]
      parameters:
        - in: path
          name: gachaId
          required: true
          schema:
            type: integer
          description: "ID of the gacha"
      responses:
        '200':
          description: "Gacha retrieved successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gacha'
        '404':
          description: "Gacha not found"
    put:
      summary: "Update a gacha by ID"
      tags: ["Gacha"]
      parameters:
        - in: path
          name: gachaId
          required: true
          schema:
            type: integer
          description: "ID of the gacha to update"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGachaRequest'
      responses:
        '200':
          description: "Gacha updated successfully"
        '400':
          description: "Invalid gacha data"
        '404':
          description: "Gacha not found"
    delete:
      summary: "Delete a gacha by ID"
      tags: ["Gacha"]
      parameters:
        - in: path
          name: gachaId
          required: true
          schema:
            type: integer
          description: "ID of the gacha to delete"
      responses:
        '200':
          description: "Gacha deleted successfully"
        '404':
          description: "Gacha not found"
  /gacha/random:
    get:
      summary: "Get a random gacha"
      tags: ["Gacha"]
      responses:
        '200':
          description: "A random gacha retrieved successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gacha'
        '404':
          description: "No gachas found" 
...
  /helloPlayer:
    get:
      summary: "Verify player token"
      tags: ["Player"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successfully verified player token"
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
        '500':
          description: "Internal server error"
  /helloAdmin:
    get:
      summary: "Verify admin token"
      tags: ["Admin"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successfully verified admin token"
          content:
            application/json:
              schema:
                type: object
                properties:
                  adminId:
                    type: integer
        '500':
          description: "Internal server error" 
...
components:  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Gacha:
      type: object
      properties:
        id:
          type: integer
          description: "Unique ID of the gacha"
        name:
          type: string
          description: "Name of the gacha"
        image:
          type: string
          description: "URL of the gacha's image"
        rarityPercent:
          type: number
          format: float
          description: "Rarity percentage of the gacha"
        description:
          type: string
          description: "Description of the gacha"
        verified:
          type: boolean
          description: "Indicates whether the gacha is verified"
    CreateGachaRequest:
      type: object
      properties:
        name:
          type: string
          description: "Name of the gacha"
        image:
          type: string
          description: "URL of the gacha's image"
        rarityPercent:
          type: number
          format: float
          description: "Rarity percentage of the gacha"
        description:
          type: string
          description: "Description of the gacha"
    UpdateGachaRequest:
      type: object
      properties:
        name:
          type: string
          description: "Updated name of the gacha"
        image:
          type: string
          description: "Updated URL of the gacha's image"
        rarityPercent:
          type: number
          format: float
          description: "Updated rarity percentage of the gacha"
        description:
          type: string
          description: "Updated description of the gacha"
        verified:
          type: boolean
          description: "Indicates whether the gacha is verified"   
...
security:
  - bearerAuth: []